requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.6.0
bs4==0.0.1
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.6
djangorestframework==3.13.1
flake8==4.0.1
idna==3.3
kombu==5.2.4
lxml==4.9.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.10
urlparse3==1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
